+++ packages/cinterop/src/commonMain/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -92,6 +92,7 @@ expect object RealmInterop {
     fun realm_freeze(liveRealm: NativePointer): NativePointer
     fun realm_is_frozen(realm: NativePointer): Boolean
     fun realm_close(realm: NativePointer)
+    fun realm_delete_files(path: String)
 
     fun realm_get_schema(realm: NativePointer): NativePointer
     fun realm_get_schema_version(realm: NativePointer): Long