+++ test/base/src/commonMain/kotlin/io/realm/test/platform/PlatformUtils.kt
@@ -1,11 +1,12 @@
 package io.realm.test.platform
 
+import io.realm.test.util.Utils
 import kotlin.time.Duration
 import kotlin.time.ExperimentalTime
 
 // Platform dependant helper methods
 expect object PlatformUtils {
-    fun createTempDir(): String
+    fun createTempDir(prefix: String = Utils.createRandomString(16)): String
     fun deleteTempDir(path: String)
     @OptIn(ExperimentalTime::class)
     fun sleep(duration: Duration)